fix(persistence): resolve sql injection and implement parameter bounds
- Replace LIMIT string interpolation with parameterized placeholders across all Turso storage modules.
- Implement MAX_TAGS_PER_QUERY bound (100) and enforce retrieval limits in tag_operations.rs.
- Refactor mod.rs to move internal embedding implementations to embeddings_internal.rs, maintaining <500 LOC per file.
- Clean up unused imports in embedding_backend.rs.

Security: Prevents potential SQL injection via LIMIT clause and resource exhaustion from unbounded tag queries.
