+++ memory-storage-turso/src/storage/mod.rs
@@ -9,9 +9,6 @@
 //! - `search`: Vector similarity search
 //! - `capacity`: Capacity-constrained storage
 
-use crate::TursoStorage;
-use do_memory_core::Result;
-
 // Re-export submodules
 pub mod batch;
 pub mod capacity;
@@ -45,38 +42,5 @@ pub use tag_operations::TagStats;
 #[cfg(feature = "turso_multi_dimension")]
 pub use embeddings_multi::DimensionStats;
 
-impl TursoStorage {
-    // ========== Backend-compatible embedding methods ==========
-
-    /// Store an embedding (backend API)
-    pub async fn store_embedding_backend(&self, id: &str, embedding: Vec<f32>) -> Result<()> {
-        self._store_embedding_internal(id, "embedding", &embedding)
-            .await
-    }
-
-    /// Get an embedding (backend API)
-    pub async fn get_embedding_backend(&self, id: &str) -> Result<Option<Vec<f32>>> {
-        self._get_embedding_internal(id, "embedding").await
-    }
-
-    /// Delete an embedding (backend API)
-    pub async fn delete_embedding_backend(&self, id: &str) -> Result<bool> {
-        self._delete_embedding_internal(id).await
-    }
-
-    /// Store embeddings in batch (backend API)
-    pub async fn store_embeddings_batch_backend(
-        &self,
-        embeddings: Vec<(String, Vec<f32>)>,
-    ) -> Result<()> {
-        self._store_embeddings_batch_internal(embeddings).await
-    }
-
-    /// Get embeddings in batch (backend API)
-    pub async fn get_embeddings_batch_backend(
-        &self,
-        ids: &[String],
-    ) -> Result<Vec<Option<Vec<f32>>>> {
-        self._get_embeddings_batch_internal(ids).await
-    }
-}
+// Methods are implemented in submodules (embedding_backend, embeddings_internal, etc.)
+// and attached to TursoStorage via impl blocks in those files.
